Add checked constructor to `PrimeField` trait
We add a new constructor `new_with_cfg` to `PrimeField` trait
whose intent is to take an arbitrary inner integer as its
argument and turn it into the Montgomery representation.

We also add an analogous constructor `new` to `ConstPrimeField` trait.

On top of that, some missing but handy const methods are added to
`MontyField`.